requirements.txt
flask==3.0.3
pandas==2.1.4
pykx==2.5.2
python-dotenv==1.0.1